test_beta/requirements.txt
APScheduler==3.8.1
argcomplete==2.0.0
attrs==21.4.0
awscli==1.22.73
backports.zoneinfo==0.2.1
bcrypt==3.2.0
boto3==1.21.18
botocore==1.24.18
certifi==2021.10.8
cffi==1.15.0
cfn-flip==1.3.0
charset-normalizer==2.0.12
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.3
cryptography==36.0.2
cycler==0.11.0
docutils==0.15.2
durationpy==0.5
ec2-metadata==2.8.0
Fiona==1.8.21
Flask==2.0.1
fonttools==4.29.1
future==0.18.2
geopandas==0.10.2
greenlet==1.1.2
gunicorn==20.1.0
hjson==3.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
kappa==0.6.0
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
munch==2.5.0
mysql-connector-python==8.0.26
networkx==2.7.1
numpy==1.22.2
osmnx==1.1.2
packaging==21.3
pandas==1.4.2
paramiko==2.7.1
pep517==0.12.0
Pillow==9.0.1
pip-tools==6.1.0
pkg_resources==0.0.0
placebo==0.9.0
protobuf==3.19.4
pyasn1==0.4.8
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.7
pyproj==3.3.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
requests==2.27.1
rsa==4.7.2
Rtree==1.0.0
s3transfer==0.5.1
scikit-learn==1.0.2
scipy==1.8.0
Shapely==1.8.1.post1
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
tqdm==4.64.0
troposphere==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.8
Werkzeug==2.0.3
wsgi-request-logger==0.4.6
zappa==0.54.1
